tools: extract abi-fetcher into a shared lib
Pulls the Etherscan-V2 provider registry, rate-limited HTTPS helpers
('httpsGet', 'rateLimitedHttpsGet', 'callExplorerWithRetry'), the
'getabi' and 'getsourcecode' wrappers, and 'isNonZeroAddress' out of
check-contract-functions.js into tools/scripts/lib/abi-fetcher.js.

Why: the same fetcher + proxy resolution is needed by sibling scripts
(e.g. a follow-up missing-permit detector). Today it lives only inside
check-contract-functions.js and cannot be reused without duplication.

Behavior change: none. The CLI surface, output, exit codes, rate limit
(~2.6 req/s), and proxy resolution rules are unchanged. The new module
takes a logger via 'setLogger(fn)'; check-contract-functions.js wires
its existing 'log' function in so verbose output is identical.
